Add files via upload
The Music Player is a simple project developed using JavaScript, CSS, and HTML. This project is an interesting project for listening to our self-made songs collection app. The user can play different kinds of music and enjoy the sound of their music. The user can click the forward and backward play button to change the songs.